Update README.md
Although this project was very short, this is a huge stepping stone for me. When I first started towards my degree I had no coding experience and had to start from the beginning and small projects were taking me hours. I made sure that the code had a specific output, with correct uses of brackets and semicolons. I think this code is simple and there is no other way to enhance it. The only thing I could format differently would be the use spaces in my cout line. During my code I had to use a lot of guidance from Zybooks. Again, I am new to coding so I had to go back to the basics to make sure my code was error free. My challenging part was figuring out where to start and making sure I had correct brackets and semicolons. This was a great project to get the basics down for my future projects. I made sure this program was readable by following all of the guidelines and making sure it was error free.